src/third_party/blink/web_tests/http/tests/reporting-observer/reporting-observer-worker.html

<!doctype html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/test-helpers.js"></script>
<script>
    const WORKER_URL
        = new URL('resources/deprecation-worker.js', location).href;
    function checkReport(report) {
        assert_equals(report.type, 'deprecation');
        assert_equals(report.url, WORKER_URL);
        assert_equals(typeof report.body.id, 'string');
        assert_equals(typeof report.body.anticipatedRemoval, 'string');
        assert_equals(typeof report.body.message, 'string');
        assert_equals(report.body.sourceFile, WORKER_URL);
        assert_equals(typeof report.body.lineNumber, 'number');
        assert_equals(typeof report.body.columnNumber, 'number');
    }

    promise_test(async (test) => {
        const worker = new Worker(WORKER_URL);
        test.add_cleanup(() => worker.terminate());

        const mc = new MessageChannel();
        worker.postMessage(mc.port2, [mc.port2]);
        test.add_cleanup(() => mc.port1.close());

        const reports = (await new Promise(r => {
            mc.port1.onmessage = r;
        })).data;

        assert_equals(reports.length, 1);
        checkReport(reports[0]);
    }, 'Deprecation reports on DedicatedWorker');

    promise_test(async (test) => {
        const worker = new SharedWorker(WORKER_URL);

        const mc = new MessageChannel();
        worker.port.start();
        worker.port.postMessage(mc.port2, [mc.port2]);
        test.add_cleanup(() => mc.port1.close());

        const reports = (await new Promise(r => {
            mc.port1.onmessage = r;
        })).data;

        assert_equals(reports.length, 1);
        checkReport(reports[0]);
    }, 'Deprecation reports on SharedWorker');

    promise_test(async (test) => {
        const SCOPE = new URL('resources/empty.html', location).pathname;
        const reg = await service_worker_unregister_and_register(test, WORKER_URL, SCOPE);
        await wait_for_state(test, reg.installing, 'activated');
        test.add_cleanup(() => reg.unregister());

        const frame = await with_iframe(SCOPE);

        const mc = new MessageChannel();
        const worker = frame.contentWindow.navigator.serviceWorker.controller;
        worker.postMessage(mc.port2, [mc.port2]);
        test.add_cleanup(() => mc.port1.close());

        const reports = (await new Promise(r => {
            mc.port1.onmessage = r;
        })).data;

        assert_equals(reports.length, 1);
        checkReport(reports[0]);
    }, 'Deprecation reports on ServiceWorker');
</script>

</html>
